+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
-## 0.9.2 (unreleased)
+## next (unreleased)
+- attachment file not found caught
+- output attached to the complex example in the README
+
+## 0.9.2
 - CLI: if nothing to do, assume parameters are a bone of an e-mail message to produce output
 - check GPG signing failed
 - fix GnuPG home